apps/bot/pyproject.toml
[project]
name = "contrarian-ai-bot"
version = "0.1.0"
description = "AI-powered cryptocurrency trading bot with contrarian sentiment analysis"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Contrarian AI Team" }
]

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlmodel>=0.0.14",
    "httpx>=0.26.0",
    "apscheduler>=3.10.0",
    "python-dotenv>=1.0.0",
    "asyncpg>=0.29.0",
    # Story 1.3: Kraken Data Ingestor
    "ccxt>=4.2.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
